+++ Magent/Views/ThreadList/SidebarModels.swift
@@ -35,3 +35,4 @@ class SidebarSection {
 
 final class SidebarSpacer {}
 final class SidebarProjectMainSpacer {}
+final class SidebarAddRepoRow {}